Clean up code
Commit UUID: fa7fd187-520f-4e30-b8ea-25876c6d3e95

+++ trunk/scripts/run_static_analyzer.sh
@@ -4,9 +4,7 @@ dummy_file=scan_build
 
 rm_cache_file()
 {
-	if [ "$cache_file" != /dev/null ]; then
-		rm -fv "$cache_file"
-	fi
+	test "$cache_file" = /dev/null || rm -fv "$cache_file"
 }
 
 do_configure()
@@ -22,9 +20,7 @@ set -eu
 # import Autoconf cache file path variable
 cache_file=/dev/null
 for f in /usr/share/config.site /usr/local/share/config.site; do
-	if [ -r $f ]; then
-		. $f
-	fi
+	test -r $f && . $f
 done
 
 if [ ! -f configure ]; then
